info
1. more detail for "info stats"
2. new command resetserverstat

+++ src/tendisplus/server/server_params.h
@@ -199,7 +199,7 @@ private:
 
 public:
     std::string bindIp = "127.0.0.1";
-    uint16_t port = 8903;
+    uint32_t port = 8903;
     std::string logLevel = "";
     std::string logDir = "./";
